fix(core): split flock helper into unix/windows files (fixes #324)
PR #1262 originally added acquireStoreLock inline in core/session.go
using golang.org/x/sys/unix.Flock, which fails to build on Windows:

  $ GOOS=windows GOARCH=amd64 go build ./core/
  core/session.go:980:17: undefined: unix.Flock
  core/session.go:980:41: undefined: unix.LOCK_EX
  core/session.go:989:12: undefined: unix.Flock
  core/session.go:989:36: undefined: unix.LOCK_UN

QA flagged this as P1 since it breaks `make release-all` for Windows
binaries. Split the helper into per-platform files following the
existing convention in the repo (core/runas_*.go,
agent/claudecode/proc_*.go, cmd/cc-connect/runas_startup_*.go):

  - core/session_lock_unix.go (//go:build !windows): unix.Flock
    advisory lock, same logic as before.
  - core/session_lock_windows.go (//go:build windows): best-effort
    no-op. Windows LockFileEx has different semantics (only blocks
    other LockFileEx callers, not all writers), so it does not
    protect against torn JSON from a non-cooperating process. The
    in-process sm.mu remains the primary defense. We log a warning
    once so operators see why cross-process concurrency is not
    guarded here.

Also rebased onto origin/main to resolve the CONFLICTING state in
core/session_test.go.

Verified cross-platform builds:
  $ go build ./core/                                 # Linux
  $ GOOS=windows GOARCH=amd64 go build ./core/       # Windows
  $ GOOS=darwin  GOARCH=arm64 go build ./core/       # macOS

All ./core/ tests pass (43.3s); TestSessionStoreLock still locks as
designed.
