+++ core/session.go
@@ -10,8 +10,6 @@ import (
 	"strings"
 	"sync"
 	"time"
-
-	"golang.org/x/sys/unix"
 )
 
 // ContinueSession is a sentinel value for AgentSessionID that tells the agent
@@ -961,32 +959,7 @@ func (sm *SessionManager) PruneEmptySessions() int {
 	return removed
 }
 
-// acquireStoreLock takes an exclusive advisory flock(2) on a sidecar
-// "<storePath>.lock" file. Returns a release function the caller must
-// defer. If the lock file cannot be created or the syscall fails, the
-// release function is a no-op and a warning is logged: the in-process
-// sm.mu is the primary defense, this is only a best-effort
-// cross-process guard.
-func acquireStoreLock(storePath string) func() {
-	if storePath == "" {
-		return func() {}
-	}
-	lockPath := storePath + ".lock"
-	f, err := os.OpenFile(lockPath, os.O_CREATE|os.O_RDWR, 0o644)
-	if err != nil {
-		slog.Warn("session: cannot create lock file, proceeding without cross-process lock", "path", lockPath, "error", err)
-		return func() {}
-	}
-	if err := unix.Flock(int(f.Fd()), unix.LOCK_EX); err != nil {
-		slog.Warn("session: flock failed, proceeding without cross-process lock", "path", lockPath, "error", err)
-		_ = f.Close()
-		return func() {}
-	}
-	return func() {
-		// Unlock is best-effort; the Close below releases the lock
-		// regardless. We still call Flock(LOCK_UN) explicitly so
-		// the lock is released before another process is unblocked.
-		_ = unix.Flock(int(f.Fd()), unix.LOCK_UN)
-		_ = f.Close()
-	}
-}
+// acquireStoreLock is implemented per-platform in session_lock_unix.go
+// (non-Windows: flock(2) advisory lock) and session_lock_windows.go
+// (Windows: best-effort no-op since Windows file locking semantics
+// differ and the in-process sm.mu remains the primary defense).
